修改分页bug

+++ src/main/java/com/lst/lc/page/Page.java
@@ -5,20 +5,23 @@ import java.util.List;
 
 /**
  * 通用分页页面
+ * 
  * @author sloriac
  * @param <T>
  */
 public class Page<T> {
-	
-	private int pageNow;
-	private int totalPageCount;
+
+	private long pageNow;
+	private long totalPageCount;
 	private boolean hasPre;
 	private boolean hasNext;
 	List<T> list = new ArrayList<T>();
+
 	public Page() {
 		super();
 	}
-	public Page(int pageNow, int totalPageCount, boolean hasPre,
+
+	public Page(long pageNow, long totalPageCount, boolean hasPre,
 			boolean hasNext, List<T> list) {
 		super();
 		this.pageNow = pageNow;
@@ -27,43 +30,53 @@ public class Page<T> {
 		this.hasNext = hasNext;
 		this.list = list;
 	}
-	public int getPageNow() {
+
+	public long getPageNow() {
 		return pageNow;
 	}
-	public void setPageNow(int pageNow) {
+
+	public void setPageNow(long pageNow) {
 		this.pageNow = pageNow;
-		if(pageNow == 1){
+		if (pageNow == 1) {
 			setHasPre(false);
-		}else{
+		} else {
 			setHasPre(true);
 		}
 	}
-	public int getTotalPageCount() {
+
+	public long getTotalPageCount() {
 		return totalPageCount;
 	}
-	public void setTotalPageCount(int totalPageCount) {
+
+	public void setTotalPageCount(long totalPageCount) {
 		this.totalPageCount = totalPageCount;
-		if(pageNow < totalPageCount){
+		if (pageNow < totalPageCount) {
 			setHasNext(true);
-		}else{
+		} else {
 			setHasNext(false);
 		}
 	}
+
 	public boolean isHasPre() {
 		return hasPre;
 	}
+
 	public void setHasPre(boolean hasPre) {
 		this.hasPre = hasPre;
 	}
+
 	public boolean isHasNext() {
 		return hasNext;
 	}
+
 	public void setHasNext(boolean hasNext) {
 		this.hasNext = hasNext;
 	}
+
 	public List<T> getList() {
 		return list;
 	}
+
 	public void setList(List<T> list) {
 		this.list = list;
 	}
